pyproject.toml
[project]
name = "efmob-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.14,<4",
  "django-environ>=0.11",
  "django-filter>=24.2",
  "drf-spectacular>=0.27",
  "bcrypt>=4.1",
  "psycopg2-binary>=2.9",
  "PyJWT>=2.8",
]

[tool.ruff]
line-length = 88
exclude = [
    "config/",
    "migrations/",
    "__pycache__/",
    ".git/",
    ".venv/",
    "venv/",
]

[tool.ruff.lint]
select = ["I", "C", "E", "F", "W"]